+++ log_test.go
@@ -1,0 +1,7 @@
+package kafkaclient
+
+import "testing"
+
+func TestErrCodes(t *testing.T) {
+
+}
