+++ message.go
@@ -10,5 +10,4 @@ type ConsumerMessage interface {
 	Offset() int64
 	Partition() int32
 	Value() []byte
-	InfoEvent(event string) string
 }
